refactor: split up map-related code
